Note the origin of the invariants (#20)
After this patch, we note down which classes and constrained
primitives the invariants are specified for. This allows us to consider
only the invariants specified for a particular class and ignore the
invariants inherited from the ancestors.

This will be useful when we generate schemas since invariants in the
schemas are inherited, unlike, say, C# code.