Remove `ExpressionWithDeclarations` from our tree
Since we do not have a clear use case yet for
`ExpressionWithDeclarations` (and consequently for `Declaration`), we
remove these two classes from *our* internal abstract syntax tree.

Additionally, there were breaking changes between Python 3.8 and 3.9
which have repercussions on parsing indices (and hence how we get to
`ExpressionWithDeclarations`). At this moment, we decide not to support
the concept and avoid the additional implementation complexity.

Once and if we have a clear use case, we will add these two classes
back into the parse tree.