Include test and CI code in pylint precommit check (#35)
We excluded the test and continuous integration code from pylint checks
since we did not really care about most of the rules. However, if
there is a ``TODO`` in the tests, there is no check to actually detect
it. There were already multiple cases where we checked in ``TODO`` code
inadvertently.

Therefore we add the test code in pylint check in this patch.

This patch also includes all the fixes which were necessary to make
pylint happy.