Upgrade docutils to 0.18.1
We have to lift the minimum docutils version to 0.18.1 since we use some
of its novel features (*e.g.*, `findall`) which are not available in the
current docutils version 0.17.1.

The problem went unnoticed since we always installed the newest version
locally, and became only apparent when the lowest version has been
installed.

Fixes #42.