Approximate naming and `allOf` to aas-specs JSON
We always introduced an `allOf` even though there is a single
definition. With this patch, we check for the number of possible
definitions and, if there is only one, write the properties directly in
the definition.

Additionally, we fix the naming to coincide more with the official JSON
schema in the http://github.com/admin-shell-io/aas-specs.