In-line constrained primitives in JSON Schema (#77)
Since the JSON schema is going to be also used for OpenAPI 3 schema, we
should not introduce additional definitions. Therefore we in-line the
constraints of constrained primitives in the actual properties of the
objects.

This change also had repercussions on SHACL generation as we in-line
the constrained primitives there as well.