Document the bottleneck in C# verification (#88)
We add a remark to warn the users of the code that passing paths around
is expensive. Hopefully we will be notified when this becomes a problem.
Otherwise, if the time permits at the end we plan to come back and
optimize the error handling so that we avoid this cost.