Optimize path handling in C# JSON deserialization (#90)
Currently, we pass in the path on every call to a deserialization
method. This is problematic since this induces a time complexity
of `O(n^2)` with `n` being the maximum length of a path. We need to
append the suffix on each call and therefore make a complete copy of the
previous path (the strings are stored as blocks in memory akin to arrays
in C#).

With this patch, we use an intermediate structure (`Error`) which stores
the path as a linked list. We append to this linked list when we unwind
the calls so that the time complexity is brought down to `O(n)`.

Additionally, the suffixes are appended only in case of errors which
should yield even better performance assuming that the correct data is
usually much more probable than incorrect data.