Refactor verification in `IEnumerable` in C# (#93)
Currently, we use `Errors` as a custom-tailored container. In this
patch, we use `IEnumerable` and `yield return` instead to use standards
language idioms.

This also allows us to reduce the time complexity to `O(n)` (from
`O(n^2)`) where `n` is the length of the maximum path segment.

Additionally, we substantially simplify the code to use pattern matching
instead of a powerfule and complex, but at the moment unnecessary,
unrolling.