Extract structure information from docstrings
Currently, we implemented the parsing of the docstrings using duct tape
in C# code generation. This does not scale to other languages or
browsable representation.

Thus we extract more detailed and better structured information from
docstrings in a principled manner at the intermediate stage so that the
generators become simpler and easier to write.