Exclude abstract classes from `stringify`
There is no point of enforcing abstract classes to be stringified. Hence
we exclude them from the `Dumpable` union and list only the concrete
classes. This makes the stringify modules easier to follow.

Mind that you can still write a stringification method for an abstract
class -- it is simply not enforced to have such a method.