Simplify iteration over intermediate doc elements
We used two functions (`_over_descriptions` and
`_over_symbol_descriptions`) in `intermediate._translate` only to call
`findall` on the corresponding documents.

In this patch, we integrate the `findall` in a single function,
`_find_all_in_descriptions`. This saves us one indention level and makes
the code even easier to follow.