Script syncing of the test meta-models (#113)
We group the test meta-models by version and enforce that they coincide
within the same group. However, we often develop the code and *change*
a meta-model at the same time. Once we are done with the development, we
had to tediously sync the meta-models within the group by hand.

In this change, we write a script to automate this syncing in order to
speed the development process.