Return errors instead of raising in C# jsonization (#114)
We had an assertion before warning against a missing feature. This is
not informative for the end user since it seems as if there is
a program bug rather than a missing feature.

With this change, we return an error which is then accompanied by the
line and column for the end user, as they might expect.