Remark why pre-conditions are not conjunctions
We add a note in the types of the intermediate representation why the
pre-conditions of a method are not modeled in the disjunctive normal
form which is common for programming languages supporting contracts.

Namely, we do not handle overriding of the methods at the moment, so we
leave this change in the intermediate representation for the future. The
note was however necessary to avoid confusion.