Test with ``aas-core-meta`` releases (#138)
Previously, we manually copy/pasted ``meta_model.py`` from
``aas-core-meta`` repository. While this was simple, it became tedious
and intractable as the number of changes increased.

Instead, we started publishing ``aas-core-meta`` releases and using them
directly in this repository as a development dependency. This makes it
easier for us to spot exactly which version of the meta-model we are
testing against.