Verify the limitedness of type annotations
We check at the intermediate stage that only a limited set of type
annotation patterns is used. This greately simplifies the generators.

However, we anticipate that more complex type patterns are going to be
used in the meta-models in the future. At the moment, we stick to the
simple generators, and explicitly check for these simplifying
assumptions.