Remove ``textwrap.dedent`` (#163)
We used ``textwrap.dedent`` thinking that it would help us with the code
layout. However, there were too many bugs when we included snippets
*within* the text which included newlines and indention of their own.

Spotting and fixing these bugs was far too tedious, so we decided to
remove ``textwrap.dedent`` from the code base altogether as we think
that the benefits (better code layout) are not worth the drawbacks (hard
to fix bugs).