Make stringify output multi-line string
Currently, stringify outputs a single-line string representation of
text. This makes it very hard to diff test data against changes when we
manually need to inspect it.

Hence we render the multi-line text in the stringified test data as
multi-line string literals to make the diffing easier for the developer.