Fix stacking of inferred schema constraints (#166)
We infer the schema constraints (currently only length and patterns)
based on the meta-model. We do not inherit them by default since we
usually generate the invariants for each class in the schema
*separately* and let the schema engine deal with the inheritance.

However, in the case of test data generation, it would be helpful if the
schema constraints are properly inherited among the classes.

In this patch, we re-structure the code and the tests to cater for that
use case and extend the logic with an additional function to stack the
schema constraints by the class ontology.

Unsurprisingly, a couple of bugs were discovered and fixed along
the re-structuring.