Add experimental support for ``any`` in invariants (#173)
We did not thouroughly test the transpilation of ``any`` in invariants,
but add it to our parsing. This is necessary since we need to create a
new version of aas-core-meta which uses ``any`` in some of the
invariants and the smoke tests fail at the moment.

Non-experimental support and testing is coming in one of the upcoming
commits.