Move ``requirements-dev`` back to ``setup.py``
We have to undo #160 since it actually broke the packaging. The issue
went unnoticed as we haven't installed the new version from pypi.org
recently.

Specifically, we would have to package ``requirements-dev.txt`` with the
package so that ``setup.py`` can read it on installation. It makes
little sense to package development requirements for installation,
though.