Fix XSD for abstract classes without implementers
In cases where an abstract class is defined in the meta-model, but no
implementers are provided, we generated an invalid XSD. Namely, we
expected a choice group to exist, but since there were no implementers,
no choice group has been generated.

In this patch, we generate an empty element for the abstract class and
simply refer to it instead of the choice.

We manually tested against aas-core-meta 2022.5.26. Therefore we also
update the example to comply with this newer meta-model despite the fact
that the repository still uses aas-core-meta 2022.5.30a6 for testing.

We will update to the latest aas-core-meta 2022.5.26 very soon in one of
the upcoming commits.