Fix duplicate inheritance of pattern constraints
In cases where a property is a constrained primitive, the stacking of
pattern constraint over ancestor is broken, as we inherit the
constraints multiple times.

As we do not keep track of where the constraints of the ancestors come
from, we simply de-duplicate by keeping a set of observed patterns.