Infer non-nullness in the intermediate
We infer the non-nullness by agressively making assumptions in
conjunctions and implications on "is not None"'s.

Word of caution. This might be too aggressive since we assume that
function calls and method calls do not alter the values observed in the
implications or conjunctions. This indeed holds true as long as you
observe query-command separation, but only that long.