Rename "symbol" to "our type" (#222)
We rename the concept related to "symbol" to "our type". That is because
we plan to introduce constants in the symbol table. Therefore, "symbol"
is a misnomer as all constants, verification functions and types are
"symbols".

As "our type" is already used in type annotations
(``OurTypeAnnotation``), we renamed all ``symbol`` to ``our_type``.

This renaming produced some awkward names down the line ("our type
reference") so we also renamed those classes and variables (*e.g.*,
"reference to our type").