Update logic for min. length of 1 for lists
We update for [aas-core-meta 08f5bb1], which requires that
all lists are either null or at least have a length 1 (except for a few
exceptions).

We therefore have to consider a special case in SHACL schema generation
since SHACL does not distinguish between empty and null properties.

Finally, we update the test data to reflect [aas-core-meta 08f5bb1].

[aas-core-meta 08f5bb1]: https://github.com/aas-core-works/aas-core-meta/commit/08f5bb1