Update to date specification invariants
Transpiling data specification invariants revelead a bug related to
constant sets in C#. Namely, the hash sets should operate on nullables
instead of non-nullables, as the former are references and the latter
are values in C#. We fix this in this change.

We also update the test data to [aas-core-meta 95e7dbf] which directly
make use of this change.

[aas-core-meta 95e7dbf]: https://github.com/aas-core-works/aas-core-meta/commit/95e7dbf