Create test empty snippets for RDF+SHACL (#256)
Currently, RDF+SHACL generation does not depend on any snippets, which
will probably change in the near future. Therefore, we keep the logic
for reading the snippets from a directory, and create an empty snippets
directory in the test data.

We mistakenly broke the tests since Git automatically removed empty
directories.