Update test data to aas-core-meta 1df1729
We update the development requirements to and re-record the test data
for [aas-core-meta 1df1729].

So far, we referenced the properties as properties of a concrete class
in the documentation instead of referencing them to the most abstract
class which defines them.

This creates problems with [docfx] in C# so we reference the
properties as properties of the abstract class.

We also update the implementation-specific snippets for C# which also
referenced the properties in the concrete class instead of the abstract
one.

[aas-core-meta 1df1729]: https://github.com/aas-core-works/aas-core-meta/commit/1df1729
[docfx]: https://dotnet.github.io/docfx/