Hard-wire ``Lang_string`` in RDF+SHACL (#263)
We hard-wire the handling of the class ``Lang_string`` in RDF+SHACL, and
represent it as the native ``rdf:langString``.

Admittedly, this is hacky. We could have made the class ``Lang_string``
implementation-specific and defined its ``rdfs:range`` manually as
a snippet.

However, we decided against that as such a design would force us to
define ``langString`` for every language and schema which do not
natively support it, write custom data generation methods *etc.* Given
that RDF+SHACL codegen is one out of many code generators we leave the
other code generators and test data generators as simple as possible,
and make the RDF+SHACL code generator a bit hacky in return.