Hard-wirte `Value_data_type` to `xs:anySimpleType`
In XSD, RDF and SHACL, ``Value_data_type`` should be represented as
``xs:anySimpleType`` instead of a string. We hard-wire how we represent it
in XSD and RDF+SHACL schema generators.

Similar to ``Lang_string``, this hard-wiring is hacky. We could have made
the class ``Value_data_type`` implementation-specific and defined it
manually as a snippet.

However, we decided against that. This would be a major hurdle for
other code and test data generators (which can treat ``Value_data_type``
simply as string). Therefore, we make the XML and RDF+SHACL schema
generators a bit more hacky instead of complicating the other
generators.

If in the future, for whatever reason, the semantic of ``Value_data_type``
changes (or the type is renamed), be careful to maintain backwards
compatibility! You probably want to distinguish different versions
of the meta-model and act accordingly. At that point, it might also make
sense to refactor these schema generators to separate repositories, and
fix them to a particular range of meta-model versions.

This change addresses the following issue:
https://github.com/admin-shell-io/aas-specs/issues/210