Restrict value data types in XML (#269)
We hard-code the member types which are allowed as `Value_data_type`.
This is really hacky -- as we iterate over concrete enumeration from the
model. If the model changes, and provides no ``Data_type_def_XSD`` or
``Value_data_type`` changes the semantics *etc.*, the generator for XML
schema is messed up.

Yet, we deliberately decided to keep other generators simple, and make
XML generator hacky for this particular pain point.

If in the future, for whatever reason, the semantic of
``Value_data_type`` changes (or the type is renamed), be careful to
maintain backwards compatibility here! You probably want to
distinguish different versions of the meta-model and act accordingly.
At that point, it might also make sense to refactor this schema
generator to a separate repository, and fix it to a particular range of
meta-model versions.