Restrict `Value_data_type` to `xs:string` in XSD (#270)
We hard-wire the constrained primitive `Value_data_type` to a simple
type with `xs:string` restriction in the XSD generator.

Previously, we hard-wired it to a list corresponding to
`Data_type_def_XSD`. This list, however, is only compatible with XSD 1.1
and breaks XSD 1.0. Since the effect of the restriction is identical, we
opt for compatibility with XSD 1.0.

See also the important note in the commit [5ae9681] regarding how hacky
this whole hard-wiring is. In the near future, you should re-consider to
refactor the XSD generator out of this repository and pin it to a range
of AAS meta-model versions.

[5ae9681]: https://github.com/aas-core-works/aas-core-codegen/commit/5ae9681632477852bda9b7ced18730a22b6492bd