Make remote CI work again (#283)
We only considered excludes in the ``--overwrite`` mode of the
pre-commit checks, but forgot to include them also in the non-invasive
checks as well. This patch fixes the issue, and the CI runs again.

A minor pylint issue (unused import) had not been detected before, but
now gets detected on the remote server, so we also remove this unused
import.

Additionally, locally we installed the new aas-core-meta version, but we
mistakenly did not update it in setup.py. In this patch, we update to
the latest aas-core-meta version 3ffe727.