Fix validation of XML types in Python
Due to a glitch in the unit test code of the Python SDK, we did not test
the handling of XML values properly. The examples were not globbed
recursively, but only at the top-level example directory.

In this patch, we make a couple of minor changes which now fixes the
test cases so that the unit tests pass:

* All matching functions (``matches_*``) return a boolean, but we
  checked for ``None``, which basically invalidated the checks.
* We use a short regular expression to parse dates. The previous code
  based on ``.split('-')`` was too error prone.
* We did not properly check for infinity in ``xs:float``'s. After this
  change, we check for inifinity just like for ``xs::double``.