Sync naming with the published schemas for V3RC02
Unfortunately, we had a bug in aas-core-codegen e2a7a806 (2022-10-30)
and prior versions where we generated the names for JSON and RDF
schemas based on a hard-wired list of abbreviations in
*aas-core-codegen* instead of relying on the naming in aas-core-meta.
This list has not been updated before publishing the schemas, so we
have to stick with the unexpected casing in the names. To maintain
the compatibility with the schemas, we allow for capitalization of
certain abbreviations where uppercase would be expected. Concretely,
this was the case with abbreviations "AAS" and "XSD".

We remove the hard-wired list in aas-core-codegen (this repository).
Additionally, we re-visit the naming algorithms and make them more
robust towards the input from aas-core-meta.

To regenerate the test data accordingly, we update to
[aas-core-meta d180dbb], which has the naming fixed.

[aas-core-meta d180dbb]: https://github.com/aas-core-works/aas-core-meta/commit/d180dbb