Report errors if classes missing for RDF and XSD
We hard-wire ``Lang_string`` and ``Value_data_type`` in RDF and XSD
generators, respectively. This is indeed hacky (see the explanations in
the patch), but this way we keep the other code and test data
generators much simpler.

Previously, if the expected types ``Lang_string`` and
``Value_data_type`` have not been defined in the meta-model, we raised
exceptions which confused the user. In this patch, we report informative
errors explaining the background to the user as well.