Make Regex renderer public (#295)
We refactor the rendering in ``parse.retree`` in such a way that
generators can implement their own rendering. For example, in TypeScript
you have to write ``\u{...}`` while in Golang you write ``\x{...}`` for
unicode characters, which differs from Python ``\u...`` and ``\U...``.

Additionally, we make a minor change to precommit script to set pylint
CWD to the repository root so that we can run pre-commit checks from
other directories as well.