De-duplicate code in ``naming`` (#298)
We refactor the code in the ``naming`` module into fewer functions, so
that the dependent naming modules can re-use them. This makes it more
obvious for the reader which casing is used.

While duplication is in similar cases often better for readability,
we already observed related bugs in the TypeScript generator (which
is still a work-in-progress), so we react accordingly.