Generate empty constructors in C#
While we do not test for this logic at the moment, we reviewed the C#
code while writing the TypeScript generator (which is still
work-in-progress). It turns out that we missed the case where there are
statements in C# constructors, but no arguments.

In this patch, we generate the constructors as soon as they have
statements, regardless of the number of arguments.