Generate constructors with no arguments in Python (#304)
While we do not test for this logic at the moment, we reviewed
the Python code while writing the TypeScript generator (which is still
work-in-progress). It turns out that we missed the case where there are
statements in constructors, but no arguments.

In this patch, we generate the constructors as soon as they have
statements, regardless of the number of arguments (which can be zero or
more).