Fix transpilation of function names in Python (#306)
We simply transpiled the function name to Python casing in
transpilation. While this works for the meta-models designed so far, it
is a ticking bomb to explode if we have cross-module function calls.

This patch fixes the issue by delegating the transpilation of the name
to the concrete implementation in the abstract transpiler class.