Fix RDF+SHACL when ``Lang_string`` is missing
We assumed that ``Lang_string`` will always be defined, and hard-wired
certain parts of the RDF+SHACL codegen to that particular class.

However, this assumption did not hold long: already the first iteration
of V3.0 of the metamodel came without a ``Lang_string``.

In this patch, we make RDF+SHACL more robust in face of missing
``Lang_string``, and allow the generation without its definition.