Refine error on unexpected SHACL len constraints
When a length constraint has been inferred for a property of type class,
the SHACL generator returned a rather confusing, or even misleading,
error message.

Though a length constraint on a property of type class is probably a
mistake in the meta-model, it is a probable mistake and deserves a more
specific error message.

This patch refines the error messages and gives more precise information
to the user so that the error can be localized swiftly.