Transpile everything to C# in smoke mode (#326)
We encountered various errors during the design of the aas-core-meta V3
which would have been prevented if only we transpiled to C# (or any
other language). Notably, we needed type inference which was missing in
the current smoke tests.

This patch mocks all the implementation specific snippets and transpiles
to C# in the smoke command, so that the errors in the meta-models can be
discovered earlier during the design.