Change JSON naming of model types for V3
The current naming algorithm for model types turned out to be rather
fragile. Namely, we used CamelCase except for abbreviations, which were
left as-are. This was brittle: if we forgot to uppercase the
abbreviation in the meta-model, and published the JSON schema, there was
no way to put the Gini back into the bottle.

Instead, in V3, we introduce a simpler, but also a more robust
algorithm: we simply CamelCase everything, including the abbreviations.