Allow strengthening of types (#365)
We allow classes to inherit properties from parent classes and
strengthen them in terms of behavioral subtyping. Namely, a property in
the class can be non-null version of the parent property, a more
concrete version of the parent property, or both.

This allows us to model some model constraints in the type system
instead of invariants. This, in its turn, allows us to generate schemas
which enforce these constraints or detect inconsistencies at compile
time.