Add `descendants` property to `Class` (#376)
Currently, we only track concrete descendants of a class. This works
well for transpilation, but it was insufficient for visual
representations such as [htmlgen in aas-core-meta]. Namely, the readers
were much more interested in *all* descendants, while the concrete
descendants were secondary.

Therefore, we add `descendants` to `Class` to list all descendants of
a class, and not only the concrete ones.

[htmlgen in aas-core-meta]: https://github.com/aas-core-works/aas-core-meta/tree/main/htmlgen