Revert strengthening of types (#379)
We revert the strengthening of types in the children classes (#365) as
it turned out that:

1) Strengthening was not necessary for V3 (we misinterpreted Constraints
   AASd-117), and
2) Strengthening is counter-intuitive. We would have had to make the
   setter hidden in the interfaces, and available only in the concrete
   classes. This, on its other hand, is tedious to implement in the code
   generator. For example, C# SDK follows an interface-all approach, so
   making the setters in the concrete classes visible is far from
   trivial.

   Current implementation used runtime checks in the setters, which are
   an incredibly dangerous footgun. The callers would probably more
   often than not forget to think about the setters doing anything
   dubious, and certainly not expect an exception there.

In the end, we decide to embrace the runtime checks through invariants
should the need for strengthening of types, and in particular
strengthening of non-nullability, arise. This will give us runtime
errors instead of exceptions, which can be gracefully handled by
the caller code and is something expected rather than unexpected.