Check for keywords in many programming languages
Prompted by the implementation of C++ SDK, we decided to check for
conflicts with reserved keywords in many different programming languages
(Ada, C, C++ *etc.*). Though we might, or might not, implement the
generators for all of them, we check for keywords in advance to detect
issues with the meta-model ahead of time.