Simplify iterations over our types using lists
We pre-compute lists for classes in #406, so we simplify the code in
this patch to make it more correct by design and more readable.

However, we made no changes to the code which would change the order of
iteration over the symbols, and consequently changing the test data.
This would be too many changes to the downstream code which we did not
consider worthwhile just for the sake of code consistency.