Fix error message in C# xmlization on declarations
When we encountered XML declarations, we threw exceptions with
uninformative messages. Namely, we expect the reader to be moved to the
content, but most users omitted to read that in the documentation.

We refine the error message for this particular situation, and hint at
`MoveToContent` method on the reader.

See [this aas-core3.0-csharp issue] for more details.

[this aas-core3.0-csharp issue]: https://github.com/aas-core-works/aas-core3.0-csharp/issues/24