Remove redundant initilaizations with None
I (mristin) was not aware of the syntax such as `name: str` where mypy
forces name to be initialized in all the execution paths. Hence, I
previously always wrote:

```python
name = None  # type: Optional[str]

...

assert name is not None
```

which is unnecessarily verbose and includes a redundant is-not-None
check at the end.