Report running tests in update-to-meta script
It turned out that a test script blocked in the update-to-aas-core-meta
development script, but the current version of the script only listed
the number of running scripts, and not which script was blocking.

In this patch, we explicitly lists which tests are still running after
the initial expected runtime, so that the developer can take action if
needed.