Fix code generation for passing of Optionals
In Java it is custom to use optionals only as return values, in all other places
we use nullable references directly. This means we have to unpack optionals for
method and function calls before passing them and provide a fallback for the
case where these optionals may unset.

Add a special case for when optionals are passed as function arguments and
convert them to null if they are unset.