Revert #482 due to a misunderstanding (#483)
The authors added in #482 supplied no commits. While they did contribute
to the project with ideas and suggestions, they did not contribute any
changes to the code or content of this repository. Since `AUTHORS`
concerns the copyright of the content, and not the acknowledgments, we
revert the changes in #482.

At this point, we do warmly acknowledge their contributions in ideas
and discussions!