Remove enum MessageType in ProtoBuf (#484)
The ProtoBuf generator introduced an additonal enum called
`MessageType`. It was intended to be used as identification of
the concrete type of message that is used in a proto3 `oneof` construct.
However, when compiling `*.proto` files, some identification mechanism
is automatically introduced by the proto compiler. Thus, this enum is
not required anymore.