Fix escaping in JSON schema patterns
The escaping logic for patterns in JSON schema got convoluted, and we
introduced inadvertently inconcistencies in the formatting. Namely, we
first *unescaped* all the characters before passing them into the parser
for regular expressions. This unescaped characters later propagated and
mixed with *escaped* characters when the expression was transpiled for
UTF-16.

To solve this mess, we remove the unescaping step at the beginning and
test more thouroughly to specify & observe the expected behavior.