Fix over-consuming XML de-serialization in Golang
We read a token too much in `unmarshalXxx` functions. The bug went
unnoticed as all our examples used indented XML, which eclipsed the
over-consumption.

In this patch, we do not generate the `unmarshalXxx`, as they turn out
to be redundant, and use `readXxxWithLookahead` to appropriately read
the tokens without over-consumption.