Verify that patterns are anchored (#499)
As we transpile to some of the systems which always expect the regular
expressions to be anchored at the start and the end (*e.g.*, XSD
engines), we verify in the patch that all verification patterns are
appropriately anchored. This ensures that the patterns will also be
correctly transpiled down the line.