Update test data to aas-core-meta bd56058
We update the development requirements to and re-record the test data
for [aas-core-meta bd56058].

Notably, we propagate the fix for AASd-109 in V3.0 where we change
the invariant such that it checks for consistency among properties even
when a ``value`` property is missing. This affects only the re-recording
of the test data corresponding to V3.0 meta-model.

[aas-core-meta bd56058]: https://github.com/aas-core-works/aas-core-meta/commit/bd56058