Check for optional type in nullness checks (#522)
We forgot to check that the value type is an optional type in type
inference. This causes various compilation errors in some of the
strongly typed languages as you can not compare non-null types with
null (*e.g.*, in C++ or Go).

In this patch, we fix the type inference to correctly detect the invalid
(non-)nullness checks and report them.