Refactor type inference in functions (#523)
So far, we were unsure what the best abstraction for type inference
would be. After a couple of code generators, the structure for the type
inference crystalized out: we refactor it into two functions, one for
the class invariants and one for the transpilable verification
functions.

The type inference for pattern verification functions is explicitly left
out as we can not infer the type of the matching result in any
meaningful way with the simple types we have so far, and we also do not
need to. Namely, we ensure that pattern verification functions follow a
very strict structure where each variable can only be a string in the
parsing stage.