Fix the order of fields in Protobuf (#530)
Previously, we relied on the order of ``set`` for the fields in
Protobuf. This is dangerous as the order of ``set`` is not well-defined
across different Python versions and implementations.

To avoid potential bugs in the future, we fix the fields to correspond
with the order of properties in the class.

Note that this solution is also not completely future proof. If you add
properties in the middle of a class, the field IDs in protocol buffers
will change causing incompatibility between the two Protocol Buffer
definitions.