Add choices for concrete classes in Protobuf (#531)
We add additional choice classes corresponding to concrete classes with
concrete descendants in Protobuf.

So far, we omitted this case as the concrete classes with concrete
descendants do not appear in the AAS meta-models. However, we decide to
make the implementation future-proof in that regard as it is very possible
that in the near future we do face concrete classes with descendants in
properties.