Use `_choice` suffix consistently in Protobuf (#534)
In #531 the `_choice` suffix has been introduced for concrete types with
concrete classes as subtypes. Since Protobuf does not support 
inheritance, this was necessary so that we can distinguish between
the concrete class and a choice (union) class with quasi-polymorphism. 
For the sake of consistency, this suffix shall be used for all types 
that have concrete descendants, *i.e.* for all the choice (union) 
classes, including the abstract classes.