Enforce None default for optional constructor args
We enforce that all optional constructor arguments must be given a
``None`` default value for consistency. We struggled in downstream
projects where optional arguments *without* a default caused a lot of
confusion.

Therefore, to make the life of the downstream users simpler (*e.g.*,
people using the generated SDKs), we impose that all constructor
arguments must be initialized with a ``None`` default, which is in line
with expectations of most of the programmers.