Allow for custom pattern fixing in jsonschema
So far, we always translated the patterns in JSON schema to surrogate
pairs, where characters above the Basic Multilingual Plane are encoded
with two characters.

This does not work for some downstream clients whose JSON Schema engine
requires a different pattern. For example, using `jsonschema` library in
`aas-core3.0-testgen` can not translate the surrogate pairs back to
Unicode characters in patterns, and thus fails on actually valid JSON
schema.

In this change, we make the `generate` function public and allow the
client to supply their own function for how the patterns should be
translated (*i.e.*, "fixed").