Handle list of primitives in JSON Schema, Python (#547)
This change makes it possible to handle list of primitives in order to
generate definitions in JSON schema as well as the de/serialization
code in Python.

This is the first change that includes integration tests to test
the pipeline end-to-end: from code generation to testing of
the *generated* code isolated from AAS or any other specific meta-model.
We included two targets (JSON schema and Python) in one pull request
so that we can specifically tailor the integration tests for these
two different use cases.