Remove stale expected file for Java
In [the fix #334 on aas-core-meta], a member of Embedded_data_specifications was
marked as non-optional. Thus, Embedded_data_specifications no longer has
optional members and consequently doesn't need a builder type anymore. Thus, the
corresponding builder file that was part of the expected output is stale. We
therefore remove the builder.