Remove stale expected file for Java (#554)
In [the fix #334 on aas-core-meta], a member of
`Embedded_data_specifications` was marked as non-optional. Thus,
`Embedded_data_specifications` no longer has optional members and
consequently doesn't need a builder type anymore. The corresponding
builder file that was part of the expected output is stale. We therefore
remove the builder.

[the fix #334 on aas-core-meta]: https://github.com/aas-core-works/aas-core-meta/pull/334