Add base64url encoding to TypeScript
The base64url encoding is heavily used on the server side, and most
users of the TypeScript SDK use it for interactions with AAS servers.
Hence, we include base64url encoding in the `common` module of the
TypeScript SDK to make their life easier even though it is not directly
used in the SDK itself.

This addresses the issue:
https://github.com/aas-core-works/aas-core3.0-typescript/issues/36