Remove dependency on aas-core-meta repository
We remove the dev. dependency on aas-core-meta repository since we want
to refactor `setup.py` into `pyproject.toml`. In this patch, we download
the latest aas-core-meta v3.py meta-model to test data, so that there is
no direct link between the packages aas-core-meta and aas-core-codegen.

This has an additional benefit that we can include fixes in
aas-core-codegen which might not be immediately applied in
aas-core-meta. Aas-core-meta uses aas-core-codegen for continuous
integration, so changes in aas-core-codegen might break aas-core-meta
and vice-versa. This already caused a couple of chicken-and-egg problems
when updating the meta-models with novel features.