Add `is_literal_of` method to symbol table
We need a method in the symbol table that allows us to succinctly check
whether a literal belongs to an enumeration or a constant set of
enumeration literals. This is practical for downstream clients which
need a lot of runtime assertions to hint the reader the expected origin
of a literal.

For example, we use pre-conditions in aas-core3.1-testdatagen to suggest
the client what inputs are expected. Writing these pre-conditions
without such a method would be quite unreadable and unmaintainable due
to a lot of redundant code.