Add literal value set to IR (#566)
For all the literal containers (enumerations, constant set of primitives
and constant set of enumeration literals), we add the property
`literal_value_set` as a frozen immutable set.

We use this in downstream code, *e.g.*, in aas-core3.1-testdatagen,
where we use it for checks in the fixing logic for automatically
generated instances.